fixed nil comparing test helpers
you can't really compare pointers to structs to nil, you have to cast
the pointer to have the type 'pointer to struct' and then compare. but
we can't do that here, because we don't know the type, so we try to use
regular comparison and fall back to reflect if that fails.

(and hope it doesn't break, because it might panic. so far it didn't do
that, so it's ok, probably?)

also, why didn't it occur earlier? because you *can* compare error to
nil, which was what we did...

references:
- http://stackoverflow.com/questions/13476349/check-for-nil-and-nil-interface-in-go
- http://golang.org/pkg/reflect/#Value.IsNil

test program:

```go
package main

import (
  "errors"
  "fmt"
  "reflect"
)

type Config struct{}

func isNil(val interface{}) bool {
  return val == nil
}

func isNilReflect(val interface{}) bool {
  return reflect.ValueOf(val).IsNil()
}

func main() {
  var c *Config = nil
  fmt.Println("comparing &config to nil")
  fmt.Println("direct:", c == nil)
  fmt.Println("indirect:", isNil(c))
  fmt.Println("reflect:", isNilReflect(c))

  var e error = errors.New("???")
  fmt.Println("comparing error to nil")
  fmt.Println("direct:", e == nil)
  fmt.Println("indirect:", isNil(e))
  fmt.Println("reflect:", isNilReflect(e))
}
```